requirements.txt
boardgamegeek2
pytest
requests
